@@ -22,30 +22,14 @@
     "__postinstall": "electron-builder install-app-deps"
   },
   "engines": {
-    "node": "12.16.3"
+    "node": "14.16.1",
+    "npm": ">= 7.0.0"
   },
   "main": "background.js",
   "dependencies": {
     "@ceramicnetwork/key-did-resolver": "^0.2.2",
-    "@chainsafe/bls-keygen": "^0.2.1",
     "@chenfengyuan/vue-qrcode": "^1.0.2",
     "@decent-bet/solido": "^1.1.6",
-    "@erebos/bzz-feed": "^0.13.2",
-    "@erebos/bzz-node": "^0.13.0",
-    "@erebos/hex": "^0.13.0",
-    "@erebos/keccak256": "^0.13.1",
-    "@erebos/secp256k1": "^0.10.0",
-    "@erebos/swarm-browser": "^0.13.0",
-    "@erebos/timeline": "^0.13.1",
-    "@ledgerhq/hw-app-eth": "^5.19.0",
-    "@ledgerhq/hw-transport-node-hid-singleton": "^5.19.0",
-    "@ledgerhq/hw-transport-webusb": "^5.17.0",
-    "@rsksmart/rif-storage": "^0.1.0",
-    "@stomp/stompjs": "^5.4.4",
-    "@sumsub/websdk": "^1.0.11",
-    "@swipewallet/venus-js": "0.0.4",
-    "@walletconnect/client": "^1.0.5",
-    "@walletconnect/qrcode-modal": "^1.0.5",
     "3id-did-provider": "^0.1.2",
     "axios": "^0.19.2",
     "bl": "^4.0.2",
@@ -73,6 +57,7 @@
     "key-did-provider-ed25519": "^1.0.0",
     "latest": "^0.2.0",
     "moment": "^2.26.0",
+    "moment-timezone": "^0.5.33",
     "multicodec": "^1.0.4",
     "multiformats": "^3.0.3",
     "node-forge": "^0.9.1",
@@ -101,21 +86,22 @@
     "vuetify": "^2.3.3",
     "vuex": "^3.1.2",
     "web-streams-polyfill": "^2.1.0",
-    "web3": "^1.2.6",
+    "web3": "^1.3.5",
     "xdvplatform-wallet": "^3.1.3",
     "xmlbuilder2": "^2.1.6"
   },
   "devDependencies": {
     "@mdi/font": "^5.0.45",
     "@mdi/js": "^5.0.45",
+    "@types/jest": "^26.0.23",
     "@types/pouchdb": "^6.4.0",
     "@types/pouchdb-find": "^6.3.5",
-    "@vue/cli-plugin-babel": "^4.4.6",
+    "@vue/cli-plugin-babel": "^4.5.12",
     "@vue/cli-plugin-e2e-cypress": "^4.4.6",
-    "@vue/cli-plugin-eslint": "^4.4.6",
-    "@vue/cli-plugin-typescript": "^4.4.6",
-    "@vue/cli-plugin-unit-jest": "^4.1.2",
-    "@vue/cli-service": "^4.4.6",
+    "@vue/cli-plugin-eslint": "^4.5.12",
+    "@vue/cli-plugin-typescript": "^4.5.12",
+    "@vue/cli-plugin-unit-jest": "^4.5.12",
+    "@vue/cli-service": "^4.5.12",
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^4.0.0",
     "@vue/test-utils": "1.0.0-beta.30",
@@ -132,7 +118,8 @@
     "node-sass": "^4.13.1",
     "sass": "^1.24.3",
     "sass-loader": "^8.0.0",
-    "typescript": "~3.8.3",
+    "ts-jest": "^26.5.6",
+    "typescript": "^4.2.4",
     "vee-validate": "^3.3.7",
     "vue-chartist": "^2.2.1",
     "vue-cli-plugin-i18n": "^0.6.0",
