@@ -17,7 +17,7 @@
             <v-col cols="11" xs="11">
               <v-file-input
                 prepend-icon="mdi-paperclip"
-                v-model="value"
+                v-model="files"
                 multiple
                 show-size
                 label="Files"
@@ -37,26 +37,56 @@
 
       <v-card-actions>
         <v-spacer></v-spacer>
-        <v-btn color="blue darken-1" text @click="show = false" :disabled="loading">Cancel</v-btn>
-        <v-btn color="blue darken-1" text @click="change" :disabled="loading">OK</v-btn>
+        <v-btn
+          color="blue darken-1"
+          text
+          data-cy="cancel-button"
+          @click="updateShow(false)"
+          :disabled="loading"
+          >
+            Cancel
+        </v-btn>
+        <v-btn
+          color="blue darken-1"
+          text
+          data-cy="ok-button"
+          @click="sendFiles"
+          :disabled="loading"
+          >
+            OK
+        </v-btn>
       </v-card-actions>
     </v-card>
   </v-dialog>
 </template>
+
 <script lang="ts">
-import { Component, Prop, Vue } from 'vue-property-decorator';
+import Vue from 'vue';
+import { Component, Emit, Prop } from 'vue-property-decorator';
 
 @Component({
   name: 'xdv-upload',
-  props: ['show', 'value', 'loading', 'uploadStatus'],
 })
-export default class Upload extends Vue {
-  value: string;
-  show;
-  
-  change() {
-    this.$emit('input', this.value);
-    this.show = false;
+export default class UploadDialog extends Vue {
+  @Prop()
+  readonly show: string;
+
+  @Prop()
+  readonly loading: boolean;
+
+  @Prop()
+  readonly uploadStatus: string;
+
+  files: File[] = [];
+
+  @Emit('update:show')
+  updateShow(newValue: boolean) {
+    return newValue;
+  }
+
+  @Emit('result')
+  sendFiles() {
+    return this.files;
   }
 }
 </script>
