@@ -1,4 +1,4 @@
-`<template>
+<template>
   <v-container fluid class="down-top-padding">
     <v-dialog v-model="removeDialog" max-width="500px">
       <v-card>
@@ -87,17 +87,7 @@
     <v-alert :type="alertType" v-if="alertType.length > 0">{{
       alertMessage
     }}</v-alert>
-    <v-alert
-      ><a
-        href="https://drive.google.com/file/d/1oZCZChHmedtne3E1M7wyvyQLWQm01Z1r/view?usp=sharing"
-        target="_blank"
-        >Download PKCS#11 Java Signer</a
-      ></v-alert
-    >
-    <v-card>
-      <!-- <v-toolbar color="black accent-4" dark>
-        <v-toolbar-title>Wallet</v-toolbar-title>
-      </v-toolbar> -->
+
       <v-dialog v-model="dialog" max-width="800px">
         <v-form v-model="valid" autocomplete="off">
           <v-card v-if="canCreateWallet">
@@ -263,8 +253,13 @@
           </v-card-actions>
         </v-card>
       </v-dialog>
+
       <v-expansion-panels>
         <v-expansion-panel>
+          <v-toolbar flat>
+            <v-toolbar-title>Wallet</v-toolbar-title>
+          </v-toolbar>
+
           <v-expansion-panel-header>
             <v-list-item>
               <v-list-item-avatar v-if="!existingWallet">
@@ -293,108 +288,55 @@
               </v-list-item-content>
             </v-list-item>
           </v-expansion-panel-header>
-          <v-expansion-panel-content>
-            <v-card class="mx-auto">
-              <v-card-text>
-                <v-list>
-                  <v-list-item>
-                    <v-list-item-content class="text--primary">
-                      <b>DID</b
-                      ><a
-                        target="_blank"
-                        :href="currentKeystore.walletRegistry"
-                        >{{ did._id }}</a
-                      >
-                    </v-list-item-content>
-                  </v-list-item>
 
-                  <v-list-item>
-                    <v-list-item-content class="text--primary">
-                      <b>Address</b> {{ currentAddress }}
-                    </v-list-item-content>
-                  </v-list-item>
+          <wallet-details
+            :didId="did._id"
+            :currentAddress="currentAddress"
+            :currentKeystore="currentKeystore"
+            :erc20Contract="daiContract"
+            />
+          <v-card-actions>
+            <v-spacer></v-spacer>
 
-                  <v-list-item>
-                    <v-list-item-content class="text--primary">
-                      <b>Linked to Smart Card</b>
-                      {{ getP11Name(currentKeystore) }}
-                    </v-list-item-content>
-                  </v-list-item>
-                  <v-list-item>
-                    <v-list-item-content class="text--primary">
-                      <b>Linked to P12</b>
-                      {{ getP12Name(currentKeystore) }}
-                    </v-list-item-content>
-                  </v-list-item>
-                  <v-list-item>
-                    <v-list-item-content class="text--primary">
-                      <b>Created</b> {{ new Date(currentKeystore.created) }}
-                    </v-list-item-content>
-                  </v-list-item>
-                </v-list>
-              </v-card-text>
+            <v-btn
+              v-if="currentKeystore"
+              @click="linkDialog = true"
+              text
+            >
+              <v-icon>mdi-key-link</v-icon>
+              Link External
+            </v-btn>
 
-              <v-card-actions>
-                <v-spacer></v-spacer>
+            <v-btn
+              v-if="currentKeystore"
+              @click="shareAddressDialog = true"
+              text
+            >
+              <v-icon>mdi-share</v-icon>
+              Share Addresses
+            </v-btn>
 
-                <v-tooltip top>
-                  <span>Link external</span>
-                  <template v-slot:activator="{ on }">
-                    <v-btn
-                      v-if="currentKeystore"
-                      v-on="on"
-                      @click="linkDialog = true"
-                      small
-                    >
-                      <v-icon>mdi-key-link</v-icon>
-                    </v-btn>
-                  </template></v-tooltip
-                >
+            <v-btn
+              v-if="currentKeystore"
+              @click="remove(item)"
+              text
+            >
+              <v-icon>mdi-delete</v-icon>
+              Remove
+            </v-btn>
 
-                <v-tooltip top>
-                  <span>Share address</span>
-                  <template v-slot:activator="{ on }">
-                    <v-btn
-                      v-if="currentKeystore"
-                      v-on="on"
-                      @click="shareAddressDialog = true"
-                      small
-                    >
-                      <v-icon>mdi-share</v-icon>
-                    </v-btn>
-                  </template></v-tooltip
-                >
-                <v-tooltip top>
-                  <span>Remove</span>
-                  <template v-slot:activator="{ on }">
-                    <v-btn
-                      v-if="currentKeystore"
-                      v-on="on"
-                      @click="remove(item)"
-                      small
-                    >
-                      <v-icon>mdi-delete</v-icon>
-                    </v-btn>
-                  </template></v-tooltip
-                >
-                <v-tooltip top>
-                  <span>Export wallet</span>
-                  <template v-slot:activator="{ on }">
-                    <v-btn
-                      v-if="currentKeystore"
-                      v-on="on"
-                      @click="exportWallet()"
-                      small
-                    >
-                      <v-icon>mdi-export</v-icon>
-                    </v-btn>
-                  </template></v-tooltip
-                >
-              </v-card-actions>
-            </v-card>
-          </v-expansion-panel-content>
+            <v-btn
+              v-if="currentKeystore"
+              @click="exportWallet()"
+              text
+            >
+              <v-icon>mdi-export</v-icon>
+              Export Wallet
+            </v-btn>
+          </v-card-actions>
         </v-expansion-panel>
       </v-expansion-panels>
+
       <xdv-drive
         :updateWallet="currentKeystore"
         :wallet="wallet"
@@ -406,8 +348,8 @@
         :web3="web3"
         :ethersInstance="ethersInstance"
         :ethersContract="ethersContract"
-      ></xdv-drive>
-    </v-card>
+      />
+
     <xdv-link-external-keystore
       v-model="linkExternals"
       :wallet="wallet"
@@ -417,47 +359,45 @@
     >
     </xdv-link-external-keystore>
     <xdv-unlock
-      v-model="password"
       :wallet="wallet"
+      @result="onUnlockResult"
       @load="onUnlock"
     ></xdv-unlock>
   </v-container>
 </template>
-<script lang="ts">
-import {
-  Wallet,
-  X509,
-  LDCryptoTypes,
-  DIDDocument,
-} from "xdvplatform-wallet/lib";
-import { X509Info, KeyConvert } from "xdvplatform-wallet/src/index";
-import { SwarmFeed } from "xdvplatform-wallet/src/swarm/feed";
 
-import { Component, Prop, Vue, Watch } from "vue-property-decorator";
+<script lang="ts">
+import { Wallet, X509Info } from "xdvplatform-wallet/lib";
+import { Component, Vue } from "vue-property-decorator";
 import { KeystoreIndex, DIDSigner, X509Signer } from "../shared/KeystoreIndex";
 import { ethers } from "ethers";
 import moment from "moment";
 import { Session } from "../shared/Session";
-import { pubKeyToAddress } from "@erebos/keccak256";
 import copy from "copy-to-clipboard";
-import { Subject, forkJoin } from "rxjs";
+import { forkJoin } from "rxjs";
 import Unlock from "../documents/Unlock.vue";
 import LinkExternalKeystore from "./LinkExternalKeystore.vue";
-import { it } from "ethers/wordlists";
 import Drive from "../documents/Drive.vue";
 import { DIDManager } from "./DIDManager";
 import { IPFSManager } from "./IPFSManager";
+import WalletDetails from "./WalletDetails.vue";
 import { DID } from "dids";
-import { WalletResolver } from "./WalletResolver";
 import Web3 from "web3";
-const Venus = require('@swipewallet/venus-js'); // in Node.js
 const xdvAbi = require('../../../../abi/xdv');
+
+// Extra declarations missing from `vue-google-oauth2`
+declare module "vue-property-decorator" {
+  interface Vue {
+    readonly $gAuth: any;
+  }
+}
 
 @Component({
   components: {
     "xdv-unlock": Unlock,
     "xdv-drive": Drive,
     "xdv-link-external-keystore": LinkExternalKeystore,
+    WalletDetails
   },
 })
 export default class WalletComponent extends Vue {
@@ -536,9 +476,9 @@ export default class WalletComponent extends Vue {
   ipfs: IPFSManager;
   didManager: DIDManager;
   oauthUniqueId = "";
-  web3: Web3;
-  ethersInstance: any;
-  ethersContract: any;
+  web3?: Web3 = null;
+  ethersInstance?: ethers.providers.Web3Provider = null;
+  ethersContract?: ethers.Contract = null;
   setImportSeedPhrase: boolean = false;
   importSeedPhrase: string = "";
   existingWallet: boolean = false;
@@ -554,8 +494,6 @@ export default class WalletComponent extends Vue {
     }
 
     await this.wallet.open(currentKeystore.keystore);
-    const wallet = this.wallet;
-    console.log('wallet wallet',wallet);
   }
 
   openDialog() {
@@ -564,29 +502,8 @@ export default class WalletComponent extends Vue {
     this.confirmPassword = "";
   }
 
-  getP11Name() {
-    if (
-      this.currentKeystore &&
-      this.currentKeystore.linkedExternalKeystores &&
-      this.currentKeystore.linkedExternalKeystores.pkcs11
-    ) {
-      return this.currentKeystore.linkedExternalKeystores.pkcs11.tokenIndex;
-    }
-    return "no";
-  }
-
-  getP12Name() {
-    if (
-      this.currentKeystore &&
-      this.currentKeystore.linkedExternalKeystores &&
-      this.currentKeystore.linkedExternalKeystores.pkcs12
-    ) {
-      return (
-        this.currentKeystore.linkedExternalKeystores.pkcs12.name ||
-        "No name found"
-      );
-    }
-    return "no";
+  onUnlockResult(result: string) {
+    this.password = result;
   }
 
   async onUnlock() {
@@ -613,7 +530,6 @@ export default class WalletComponent extends Vue {
     );
     /*const _this = this;
     const filter = new Promise((resolve,reject) => {
-      debugger;
       _this.contract.events.DocumentAnchored({ 
         toBlock: 'latest',
         fromBlock: 0 },
@@ -627,8 +543,6 @@ export default class WalletComponent extends Vue {
         });
     });
     const response = await filter;
-    debugger;
-    
     this.items = await forkJoin(response).toPromise();*/
   }
 
@@ -746,7 +660,6 @@ export default class WalletComponent extends Vue {
         isDefault: true,
       },
     });
-    debugger;
     await this.wallet.open(this.currentKeystore.keystore);
   }
 
@@ -828,7 +741,6 @@ export default class WalletComponent extends Vue {
     const ethersInstance = new ethers.providers.Web3Provider(web3.currentProvider as any);
     /*
     const pk = (await this.wallet.getPrivateKey("ES256K"));
-    debugger;
     const private_key = pk.getPrivate("hex");//.getSecret('hex'); 
     const account = web3.eth.accounts.privateKeyToAccount('0x'+private_key);*/
 
@@ -861,7 +773,7 @@ export default class WalletComponent extends Vue {
     this.valid = true;
 
     const wallet = new Wallet();
-    let { phrase } = Wallet.generateMnemonic() as any;
+    let { phrase } = Wallet.generateMnemonic();
     if(this.importSeedPhrase.length > 0){
       phrase = this.importSeedPhrase;
     }
@@ -880,7 +792,6 @@ export default class WalletComponent extends Vue {
         this.wallet = w;
         
         id = w.id;
-        debugger;
         keystoreIndexItem = {
           created: new Date(),
           name: this.oauthUniqueId, 
@@ -890,13 +801,12 @@ export default class WalletComponent extends Vue {
 
 
         this.currentAddress = wallet.ethersWallet.address;
-        console.log('currentAddress',this.currentAddress);
         this.alertMessage =
           "Creating DID (Decentralized Identity)...please wait";
         keystoreIndexItem.address = this.currentAddress;//pubKeyToAddress(keys.getPublic("array"));
         //this.contract = new ethers.Contract(xdvAbi.XDVDocumentAnchoring.address.bsctestnet, xdvAbi.XDVDocumentAnchoring.raw.abi, provider.getSigner());
         keystoreIndexItem.name = this.oauthName;
-        Session.set({ ks: keystoreIndexItem });
+        await Session.set({ ks: keystoreIndexItem });
       }
 
       this.loading = false;
